Create README.md
DOCKER-BLOCKBUSTER:
In this docker-compose.yml file is attached which is connected with a mysql database and have a WordPress platform.

DOCKER-COMPOSE:
It is a tool for running multi-container Docker applications. With Compose, you use a yaml file or yml file to configure your application’s services. Then, with a single command, you create and start all the services from your configuration.

STEP BY STEP GUIDE:
1. Install docker in your system, make a dockerfile environment. 
2. Define the services that make up your app in docker-compose.yml so they can be run together in an isolated environment.
3. Run docker-compose up and Compose starts and runs your entire app.